# Use PyObject_Vectorcall in rect