@ankith26 2024-08-12
I believe #3023 already covers the performance gain that can be achieved with this PR, while also using nicer API (imo).

Even if we need to support limited API in the future, it is trivial to slap in something like `#define PyObject_CallOneArg(self, obj) PyObject_Vectorcall((self), &(obj), 1, NULL)` in one of the headers